RFC: add generalized test templates, use on QEMU/grub/ipxe/uboot
Add support to add multiple test actions to an existing plan (e.g. boot)

e.g. when generating jobs, use:

  --plans boot --tests tests/test1.jinja2 tests/test2.jinja2  tests/kselftest.jinja2

This allows test templates to be independent of the boot method, and
therefore not have to have a test template for each different boot
method, since for most (all?) tests, the test should be written in a way
that should be independent of u-boot vs. grub etc.

Example job definition from the above command:
   http://lava.baylibre.com:10080/scheduler/job/923/definition

Tested with ARM64 u-boot and QEMU in lab-baylibre